+++ packages/core-cairo/src/utils/convert-strings.ts
